+++ test/math/math_power_test.exs
@@ -6,7 +6,7 @@ defmodule Math.Power.Test do
   # On my machine, iterations over 12 bring bad karma
   @iterations 12
 
-  Enum.each(1..@iterations, fn n ->
+  Enum.each(-10..@iterations, fn n ->
     test "Confirm Cldr.Math.power(5, n) for #{inspect(n)} returns the same result as :math.pow" do
       p =
         Cldr.Math.power(5, unquote(n))
@@ -24,12 +24,11 @@ defmodule Math.Power.Test do
       p =
         Cldr.Math.power(Decimal.new(5), Decimal.new(unquote(n)))
         |> Cldr.Math.round_significant(10)
-        |> Decimal.to_integer()
+        |> Decimal.to_float()
 
       q =
         :math.pow(5, unquote(n))
         |> Cldr.Math.round_significant(10)
-        |> trunc()
 
       assert p == q
     end